requirements-1.txt
opencv-python==4.13.0.92
scenedetect==0.6.7.1
numpy==2.4.2

torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
faster-whisper==1.2.1
transformers==5.2.0
pillow==12.1.1
tqdm==4.67.1

